requirements.txt
fvcore==0.1.5.post20221221
matplotlib==3.7.2
pandas==2.0.3
Pillow==11.1.0
scikit_learn==1.3.0
torch==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.6

